创建项目

+++ run.py
@@ -5,10 +5,11 @@
 
 import os
 import shutil
+from conf.setting import setting
 """
 总执行用例文件
 """
-
-shutil.rmtree("./report")
+if os.path.exists(setting.REPORT_PATH):
+    shutil.rmtree("./report")
 os.system("pytest -s -q --alluredir ./report")
 os.system("allure generate ./report/ -o ./allure-report --clean")
